+++ packages/react/src/components/CardEditor/CardEditor.story.jsx
@@ -27,8 +27,9 @@ const commonActions = {
         ]
       : [],
   dataSeriesFormActions: {
-    hideAggregationsDropDown: (editDataItem) =>
+    hasAggregationsDropDown: (editDataItem) =>
       editDataItem?.dataItemType !== 'DIMENSION' && editDataItem?.type !== 'TIMESTAMP',
+    hasDataFilterDropdown: action('hasDataFilterDropdown'),
     onAddAggregations: action('onAddAggregations'),
   },
 };