+++ packages/react/src/components/DashboardEditor/DashboardEditor.test.jsx
@@ -40,10 +40,11 @@ const commonProps = {
   actions: {
     onEditDataItem: jest.fn().mockImplementation(() => []),
     dataSeriesFormActions: {
-      hideAggregationsDropDown: jest.fn(
+      hasAggregationsDropDown: jest.fn(
         (editDataItem) =>
           editDataItem?.dataItemType !== 'DIMENSION' && editDataItem?.type !== 'TIMESTAMP'
       ),
+      hasDataFilterDropdown: jest.fn(),
       onAddAggregations: jest.fn(),
     },
   },
@@ -226,7 +227,7 @@ describe('DashboardEditor', () => {
     // there should now be two cards with the same title
     expect(screen.getAllByText('value card')).toHaveLength(2);
     // card select should have been called
-    expect(commonProps.onCardSelect).toHaveBeenCalled();
+    expect(commonProps.onCardSelect).toHaveBeenCalledWith(mockValueCard);
   });
 
   it('selecting remove card should remove card', () => {
@@ -309,7 +310,14 @@ describe('DashboardEditor', () => {
     expect(addCardBtn).toBeInTheDocument();
     fireEvent.click(addCardBtn);
     // card select should have been called
-    expect(commonProps.onCardSelect).toHaveBeenCalled();
+    expect(commonProps.onCardSelect).toHaveBeenCalledWith(
+      expect.objectContaining({
+        content: { allowNavigation: true, columns: [], showHeader: true },
+        size: 'LARGE',
+        title: 'Untitled',
+        type: 'TABLE',
+      })
+    );
   });
 
   it('selecting submit should fire onSubmit', () => {