+++ packages/react/src/components/DateTimePicker/DateTimePickerV2WithoutTimeSpinner.jsx
@@ -13,7 +13,7 @@ import {
   OrderedList,
   ListItem,
 } from 'carbon-components-react';
-import { Calendar16 } from '@carbon/icons-react';
+import { Calendar16, WarningFilled16 } from '@carbon/icons-react';
 import classnames from 'classnames';
 import { v4 as uuidv4 } from 'uuid';
 import warning from 'warning';
@@ -48,7 +48,7 @@ import {
   useRelativeDateTimeValue,
 } from './dateTimePickerUtils';
 
-const { iotPrefix } = settings;
+const { iotPrefix, prefix } = settings;
 
 export const DateTimePickerDefaultValuePropTypes = PropTypes.oneOfType([
   PropTypes.exact({
@@ -116,6 +116,8 @@ const propTypes = {
   expanded: PropTypes.bool,
   /** disable the input */
   disabled: PropTypes.bool,
+  /** specify the input in invalid state */
+  invalid: PropTypes.bool,
   /** show the relative custom range picker */
   showRelativeOption: PropTypes.bool,
   /** show the custom range link */
@@ -156,6 +158,7 @@ const propTypes = {
     hours: PropTypes.string,
     minutes: PropTypes.string,
     number: PropTypes.string,
+    invalidText: PropTypes.string,
   }),
   /** Light version  */
   light: PropTypes.bool,
@@ -213,6 +216,7 @@ const defaultProps = {
   ],
   expanded: false,
   disabled: false,
+  invalid: false,
   showRelativeOption: true,
   showCustomRangeLink: true,
   hasTimeInput: true,
@@ -243,6 +247,7 @@ const defaultProps = {
     hours: 'hours',
     minutes: 'minutes',
     number: 'number',
+    invalidText: 'The date time entered is invalid',
   },
   light: false,
   locale: 'en',
@@ -261,6 +266,7 @@ const DateTimePicker = ({
   relatives,
   expanded,
   disabled,
+  invalid,
   showRelativeOption,
   showCustomRangeLink,
   hasTimeInput,
@@ -286,7 +292,7 @@ const DateTimePicker = ({
   }, []);
 
   const langDir = useLangDirection();
-  const strings = useMemo(
+  const mergedI18n = useMemo(
     () => ({
       ...defaultProps.i18n,
       ...i18n,
@@ -309,6 +315,7 @@ const DateTimePicker = ({
   const [currentValue, setCurrentValue] = useState(null);
   const [lastAppliedValue, setLastAppliedValue] = useState(null);
   const [humanValue, setHumanValue] = useState(null);
+  const [invalidState, setInvalidState] = useState(invalid);
   const [datePickerElem, handleDatePickerRef] = useDateTimePickerRef({ id, v2: true });
   const [focusOnFirstField, setFocusOnFirstField] = useDateTimePickerFocus(datePickerElem);
   const relativeSelect = useRef(null);
@@ -402,7 +409,7 @@ const DateTimePicker = ({
       value.kind = PICKER_KINDS.PRESET;
     }
     setCurrentValue(value);
-    const parsedValue = parseValue(value, dateTimeMask, strings.toLabel);
+    const parsedValue = parseValue(value, dateTimeMask, mergedI18n.toLabel);
     setHumanValue(parsedValue.readableValue);
 
     return {
@@ -562,7 +569,7 @@ const DateTimePicker = ({
 
   const tooltipValue = renderPresetTooltipText
     ? renderPresetTooltipText(currentValue)
-    : getIntervalValue({ currentValue, strings, dateTimeMask, humanValue });
+    : getIntervalValue({ currentValue, mergedI18n, dateTimeMask, humanValue });
 
   const disableRelativeApply =
     isCustomRange &&
@@ -575,6 +582,8 @@ const DateTimePicker = ({
     (absoluteStartTimeInvalid || absoluteEndTimeInvalid);
 
   const disableApply = disableRelativeApply || disableAbsoluteApply;
+
+  useEffect(() => setInvalidState(invalid || disableApply), [invalid, disableApply]);
 
   const onApplyClick = () => {
     setIsExpanded(false);
@@ -635,7 +644,7 @@ const DateTimePicker = ({
             onClick={toggleIsCustomRange}
             onKeyUp={handleSpecificKeyDown(['Enter', ' '], toggleIsCustomRange)}
           >
-            {strings.backBtnLabel}
+            {mergedI18n.backBtnLabel}
           </Button>
         ) : (
           <Button
@@ -646,7 +655,7 @@ const DateTimePicker = ({
             {...others}
             onKeyUp={handleSpecificKeyDown(['Enter', ' '], onCancelClick)}
           >
-            {strings.cancelBtnLabel}
+            {mergedI18n.cancelBtnLabel}
           </Button>
         )}
         <Button
@@ -658,7 +667,7 @@ const DateTimePicker = ({
           size="field"
           disabled={disableApply}
         >
-          {strings.applyBtnLabel}
+          {mergedI18n.applyBtnLabel}
         </Button>
       </div>
     );
@@ -674,11 +683,13 @@ const DateTimePicker = ({
   const menuOffsetTop = menuOffset?.top ? menuOffset.top : 0;
 
   return (
-    <>
+    <div className={`${iotPrefix}--date-time-pickerv2`}>
       <div
         data-testid={testId}
         id={`${id}-${iotPrefix}--date-time-pickerv2__wrapper`}
-        className={`${iotPrefix}--date-time-pickerv2__wrapper`}
+        className={classnames(`${iotPrefix}--date-time-pickerv2__wrapper`, {
+          [`${iotPrefix}--date-time-pickerv2__wrapper--disabled`]: disabled,
+        })}
         style={{ '--wrapper-width': hasIconOnly ? '3rem' : '20rem' }}
         role="button"
         onClick={onFieldInteraction}
@@ -706,6 +717,8 @@ const DateTimePicker = ({
           className={classnames({
             [`${iotPrefix}--date-time-picker__box--full`]: !hasIconOnly,
             [`${iotPrefix}--date-time-picker__box--light`]: light,
+            [`${iotPrefix}--date-time-picker__box--invalid`]: invalidState,
+            [`${iotPrefix}--date-time-picker__box--disabled`]: disabled,
           })}
         >
           {!hasIconOnly ? (
@@ -714,13 +727,20 @@ const DateTimePicker = ({
               className={`${iotPrefix}--date-time-picker__field`}
             >
               {isExpanded || (currentValue && currentValue.kind !== PICKER_KINDS.PRESET) ? (
-                <span title={humanValue}>{humanValue}</span>
+                <span
+                  className={classnames({
+                    [`${iotPrefix}--date-time-picker__disabled`]: disabled,
+                  })}
+                  title={humanValue}
+                >
+                  {humanValue}
+                </span>
               ) : humanValue ? (
                 <TooltipDefinition
                   align="start"
                   direction="bottom"
                   tooltipText={tooltipValue}
-                  triggerClassName=""
+                  triggerClassName={disabled ? `${iotPrefix}--date-time-picker__disabled` : ''}
                 >
                   {humanValue}
                 </TooltipDefinition>
@@ -743,14 +763,19 @@ const DateTimePicker = ({
           <FlyoutMenu
             isOpen={isExpanded}
             buttonSize={hasIconOnly ? 'default' : 'small'}
-            renderIcon={Calendar16}
-            disabled={false}
+            renderIcon={invalidState ? WarningFilled16 : Calendar16}
+            disabled={disabled}
             buttonProps={{
               tooltipPosition: 'top',
               tabIndex: -1,
+              className: invalidState
+                ? `${iotPrefix}--date-time-picker--trigger-button-invalid`
+                : disabled
+                ? `${iotPrefix}--date-time-picker--trigger-button-disabled`
+                : '',
             }}
             hideTooltip
-            iconDescription={strings.calendarLabel}
+            iconDescription={mergedI18n.calendarLabel}
             passive={false}
             triggerId="test-trigger-id-2"
             light={light}
@@ -798,7 +823,7 @@ const DateTimePicker = ({
                         className={`${iotPrefix}--date-time-picker__listitem ${iotPrefix}--date-time-picker__listitem--preset ${iotPrefix}--date-time-picker__listitem--custom`}
                         tabIndex={0}
                       >
-                        {strings.customRangeLinkLabel}
+                        {mergedI18n.customRangeLinkLabel}
                       </ListItem>
                     ) : null}
                     {presets.map((preset, i) => {
@@ -818,7 +843,7 @@ const DateTimePicker = ({
                           )}
                           tabIndex={0}
                         >
-                          {strings.presetLabels[i] || preset.label}
+                          {mergedI18n.presetLabels[i] || preset.label}
                         </ListItem>
                       );
                     })}
@@ -831,7 +856,7 @@ const DateTimePicker = ({
                 >
                   {showRelativeOption ? (
                     <FormGroup
-                      legendText={strings.customRangeLabel}
+                      legendText={mergedI18n.customRangeLabel}
                       className={`${iotPrefix}--date-time-picker__menu-formgroup`}
                     >
                       <RadioButtonGroup
@@ -842,7 +867,7 @@ const DateTimePicker = ({
                         <RadioButton
                           value={PICKER_KINDS.RELATIVE}
                           id={`${id}-relative`}
-                          labelText={strings.relativeLabel}
+                          labelText={mergedI18n.relativeLabel}
                           onKeyDown={handleSpecificKeyDown(
                             ['ArrowUp', 'ArrowDown', 'ArrowLeft', 'ArrowRight'],
                             onNavigateRadioButton
@@ -851,7 +876,7 @@ const DateTimePicker = ({
                         <RadioButton
                           value={PICKER_KINDS.ABSOLUTE}
                           id={`${id}-absolute`}
-                          labelText={strings.absoluteLabel}
+                          labelText={mergedI18n.absoluteLabel}
                           onKeyDown={handleSpecificKeyDown(
                             ['ArrowUp', 'ArrowDown', 'ArrowLeft', 'ArrowRight'],
                             onNavigateRadioButton
@@ -863,13 +888,13 @@ const DateTimePicker = ({
                   {showRelativeOption && customRangeKind === PICKER_KINDS.RELATIVE ? (
                     <>
                       <FormGroup
-                        legendText={strings.lastLabel}
+                        legendText={mergedI18n.lastLabel}
                         className={`${iotPrefix}--date-time-picker__menu-formgroup`}
                       >
                         <div className={`${iotPrefix}--date-time-picker__fields-wrapper`}>
                           <NumberInput
                             id={`${id}-last-number`}
-                            invalidText={strings.invalidNumberLabel}
+                            invalidText={mergedI18n.invalidNumberLabel}
                             step={1}
                             min={0}
                             value={relativeValue ? relativeValue.lastNumber : 0}
@@ -898,7 +923,7 @@ const DateTimePicker = ({
                                 <SelectItem
                                   key={i}
                                   value={interval.value}
-                                  text={strings.intervalLabels[i] || interval.label}
+                                  text={mergedI18n.intervalLabels[i] || interval.label}
                                 />
                               );
                             })}
@@ -906,7 +931,7 @@ const DateTimePicker = ({
                         </div>
                       </FormGroup>
                       <FormGroup
-                        legendText={strings.relativeToLabel}
+                        legendText={mergedI18n.relativeToLabel}
                         className={`${iotPrefix}--date-time-picker__menu-formgroup`}
                       >
                         <div className={`${iotPrefix}--date-time-picker__fields-wrapper`}>
@@ -924,7 +949,7 @@ const DateTimePicker = ({
                                 <SelectItem
                                   key={i}
                                   value={relative.value}
-                                  text={strings.relativeLabels[i] || relative.label}
+                                  text={mergedI18n.relativeLabels[i] || relative.label}
                                 />
                               );
                             })}
@@ -982,7 +1007,7 @@ const DateTimePicker = ({
                             <TimePickerSpinner
                               id={`${id}-start-time`}
                               invalid={absoluteStartTimeInvalid}
-                              labelText={strings.startTimeLabel}
+                              labelText={mergedI18n.startTimeLabel}
                               value={absoluteValue ? absoluteValue.startTime : '00:00'}
                               i18n={i18n}
                               onChange={onAbsoluteStartTimeChange}
@@ -993,7 +1018,7 @@ const DateTimePicker = ({
                             <TimePickerSpinner
                               id={`${id}-end-time`}
                               invalid={absoluteEndTimeInvalid}
-                              labelText={strings.endTimeLabel}
+                              labelText={mergedI18n.endTimeLabel}
                               value={absoluteValue ? absoluteValue.endTime : '00:00'}
                               i18n={i18n}
                               onChange={onAbsoluteEndTimeChange}
@@ -1013,8 +1038,18 @@ const DateTimePicker = ({
             </div>
           </FlyoutMenu>
         </div>
+        {invalidState && !hasIconOnly ? (
+          <p
+            className={classnames(
+              `${prefix}--form__helper-text`,
+              `${iotPrefix}--date-time-picker__helper-text--invalid`
+            )}
+          >
+            {mergedI18n.invalidText}
+          </p>
+        ) : null}
       </div>
-    </>
+    </div>
   );
 };