+++ packages/react/src/components/HotspotEditorModal/HotspotEditorDataSourceTab/HotspotEditorDataSourceTab.test.jsx
@@ -59,10 +59,11 @@ const commonActions = {
   actions: {
     onEditDataItem: jest.fn().mockImplementation(() => []),
     dataSeriesFormActions: {
-      hideAggregationsDropDown: jest.fn(
+      hasAggregationsDropDown: jest.fn(
         (editDataItem) =>
           editDataItem?.dataItemType !== 'DIMENSION' && editDataItem?.type !== 'TIMESTAMP'
       ),
+      hasDataFilterDropdown: jest.fn(),
       onAddAggregations: jest.fn(),
     },
   },
@@ -186,10 +187,11 @@ describe('HotspotEditorDataSourceTab', () => {
           { id: 'min', text: 'Minimum' },
         ]),
         dataSeriesFormActions: {
-          hideAggregationsDropDown: jest.fn(
+          hasAggregationsDropDown: jest.fn(
             (editDataItem) =>
               editDataItem?.dataItemType !== 'DIMENSION' && editDataItem?.type !== 'TIMESTAMP'
           ),
+          hasDataFilterDropdown: jest.fn(),
           onAddAggregations: jest.fn(),
         },
       },