+++ packages/react/src/components/HotspotEditorModal/HotspotTextStyleTab/HotspotTextStyleTab.story.jsx
@@ -31,7 +31,7 @@ const commonActions = {
         ]
       : [],
   dataSeriesFormActions: {
-    hideAggregationsDropDown: (editDataItem) =>
+    hasAggregationsDropDown: (editDataItem) =>
       editDataItem?.dataItemType !== 'DIMENSION' && editDataItem?.type !== 'TIMESTAMP',
     onAddAggregations: action('onAddAggregations'),
   },